Remove unused code
Remove code that is unused either because it is substituted by other logic
or because it was ported over and never consumed.
