+++ pkg/utils/utils_test.go
@@ -2,83 +2,3 @@
 // Copyright Contributors to the Open Cluster Management project
 
 package utils
-
-import (
-	"reflect"
-	"testing"
-
-	corev1 "k8s.io/api/core/v1"
-)
-
-func TestContainsMap(t *testing.T) {
-	superset := map[string]string{
-		"hello":     "world",
-		"goodnight": "moon",
-		"yip":       "yip",
-	}
-	type args struct {
-		all      map[string]string
-		expected map[string]string
-	}
-	tests := []struct {
-		name string
-		args args
-		want bool
-	}{
-		{
-			"Superset",
-			args{
-				all:      superset,
-				expected: map[string]string{"hello": "world", "yip": "yip"},
-			},
-			true,
-		},
-		{
-			"Partial overlap",
-			args{
-				all:      superset,
-				expected: map[string]string{"hello": "world", "greetings": "traveler"},
-			},
-			false,
-		},
-		{
-			"Empty superset",
-			args{
-				all:      map[string]string{},
-				expected: map[string]string{"yip": "yip"},
-			},
-			false,
-		},
-		{
-			"Empty subset",
-			args{
-				all:      superset,
-				expected: map[string]string{},
-			},
-			true,
-		},
-		{
-			"Same keys, different values",
-			args{
-				all:      superset,
-				expected: map[string]string{"hello": "moon", "yip": "yip"},
-			},
-			false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := ContainsMap(tt.args.all, tt.args.expected); got != tt.want {
-				t.Errorf("ContainsMap() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestDistributePods(t *testing.T) {
-	t.Run("Returns pod affinity", func(t *testing.T) {
-		if got := DistributePods("app", "testapp"); reflect.TypeOf(got) != reflect.TypeOf((*corev1.Affinity)(nil)) {
-			t.Errorf("DistributePods() did not return an affinity type")
-		}
-	})
-}
